Refactored `rustcoalescence/algorithms/gillespie/src/gillespie/mod.rs`
