+++ rustcoalescence/algorithms/gillespie/src/gillespie/turnover/initialiser/mod.rs
@@ -1,0 +1,91 @@
+use necsim_core::{
+    cogs::{
+        ActiveLineageSampler, CoalescenceSampler, DispersalSampler, EmigrationExit, EventSampler,
+        ImmigrationEntry, LineageReference, LocallyCoherentLineageStore, MathsCore, RngCore,
+    },
+    reporter::Reporter,
+};
+
+use necsim_impls_no_std::cogs::{
+    dispersal_sampler::in_memory::alias::InMemoryAliasDispersalSampler,
+    origin_sampler::TrustedOriginSampler,
+};
+use necsim_partitioning_core::LocalPartition;
+
+use rustcoalescence_scenarios::Scenario;
+
+pub mod fixup;
+pub mod genesis;
+pub mod resume;
+
+#[allow(clippy::type_complexity)]
+#[allow(clippy::module_name_repetitions)]
+pub trait GillespieLineageStoreSampleInitialiser<
+    M: MathsCore,
+    G: RngCore<M>,
+    O: Scenario<M, G>,
+    Error,
+>
+{
+    type DispersalSampler: DispersalSampler<M, O::Habitat, G>;
+    type ActiveLineageSampler<
+        R: LineageReference<M, O::Habitat>,
+        S: LocallyCoherentLineageStore<M, O::Habitat, R>,
+        X: EmigrationExit<M, O::Habitat, G, R, S>,
+        C: CoalescenceSampler<M, O::Habitat, R, S>,
+        E: EventSampler<M, O::Habitat, G, R, S, X, Self::DispersalSampler, C, O::TurnoverRate, O::SpeciationProbability>,
+        I: ImmigrationEntry<M>,
+    >: ActiveLineageSampler<
+        M,
+        O::Habitat,
+        G,
+        R,
+        S,
+        X,
+        Self::DispersalSampler,
+        C,
+        O::TurnoverRate,
+        O::SpeciationProbability,
+        E,
+        I,
+    >;
+
+    fn init<
+        'h,
+        T: TrustedOriginSampler<'h, M, Habitat = O::Habitat>,
+        R: LineageReference<M, O::Habitat>,
+        S: LocallyCoherentLineageStore<M, O::Habitat, R>,
+        X: EmigrationExit<M, O::Habitat, G, R, S>,
+        C: CoalescenceSampler<M, O::Habitat, R, S>,
+        E: EventSampler<
+            M,
+            O::Habitat,
+            G,
+            R,
+            S,
+            X,
+            Self::DispersalSampler,
+            C,
+            O::TurnoverRate,
+            O::SpeciationProbability,
+        >,
+        I: ImmigrationEntry<M>,
+        Q: Reporter,
+        P: LocalPartition<Q>,
+    >(
+        self,
+        origin_sampler: T,
+        dispersal_sampler: O::DispersalSampler<InMemoryAliasDispersalSampler<M, O::Habitat, G>>,
+        turnover_rate: &O::TurnoverRate,
+        local_partition: &mut P,
+    ) -> Result<
+        (
+            S,
+            Self::DispersalSampler,
+            Self::ActiveLineageSampler<R, S, X, C, E, I>,
+        ),
+        Error,
+    >
+    where
+        O::Habitat: 'h;
+}
